refactor: Consolidate XDG test mocks into test/helpers directory
Move xdg-test.ts from test/ to test/helpers/ and remove the test/mocks/ directory to consolidate test utilities in a single location.

Changes:
- Move test/xdg-test.ts → test/helpers/xdg-test.ts
- Update all test imports from './mocks', '../mocks', './xdg-test' to './helpers/xdg-test' or '../helpers/xdg-test'
- Remove obsolete test/mocks/index.ts and test/mocks/README.md
- Fix import paths in xdg-test.ts to use ../../lib/* instead of ../lib/*

All tests continue to pass with no changes to functionality.
