--- test/mocks/index.ts
@@ -1,7 +1,0 @@
-/**
- * Test Mocks
- *
- * Centralized exports for all test mocks
- */
-
-export { XDGTest } from "../xdg-test";
